requirements.txt
certifi==2018.8.24
chardet==3.0.4
diff-match-patch==20121119
Django==1.10.8
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.6
Pillow==5.2.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
psutil
py-cpuinfo

